move files to lib folder
