+++ lib/subscriber.js
