+++ lib/subscription-connection.js
